Moves helper functions to `/utils/` drawer
