+++ app/routes/_index.jsx
@@ -2,7 +2,7 @@ import { Link } from '@remix-run/react'
 import { useLoaderData } from '@remix-run/react'
 import { OpenCarousel } from '../classes/openCarousel'
 import { renderPageLinks } from '../classes/pageHelpers'
-import { getExhibits } from '../fetch'
+import { getExhibits } from '../utils/fetch'
 
 export const loader = async () => {
   return await getExhibits()
